Move KernelStub implementation to utils/kernel_stub.rs leaving tests
This ensures every other test does not include the KernelStub testsuite as well.
