+++ plt/plt-token-module/tests/token_module_account_state.rs
@@ -1,12 +1,11 @@
-use crate::kernel_stub::{KernelStub, TokenInitTestParams};
 use concordium_base::base::ProtocolVersion;
 use concordium_base::common::cbor;
 use concordium_base::protocol_level_tokens::TokenModuleAccountState;
 use plt_scheduler_interface::token_kernel_interface::TokenKernelQueries;
 use plt_token_module::token_module;
+use utils::kernel_stub::{KernelStub, TokenInitTestParams};
 
-mod kernel_stub;
-pub mod utils;
+mod utils;
 
 /// Default protocol version used across the tests.
 const PROTOCOL_VERSION: ProtocolVersion = utils::LATEST_PROTOCOL_VERSION;
