+++ plt/plt-token-module/tests/token_module_burn.rs
@@ -1,4 +1,3 @@
-use crate::kernel_stub::{TokenInitTestParams, TransactionExecutionTestImpl};
 use assert_matches::assert_matches;
 use concordium_base::base::ProtocolVersion;
 use concordium_base::common::cbor;
@@ -7,12 +6,11 @@ use concordium_base::protocol_level_tokens::{
     RawCbor, TokenAmount, TokenBalanceInsufficientRejectReason, TokenModuleRejectReason,
     TokenOperation, TokenSupplyUpdateDetails, UnsupportedOperationRejectReason,
 };
-use kernel_stub::KernelStub;
 use plt_scheduler_interface::token_kernel_interface::TokenKernelQueries;
 use plt_scheduler_types::types::tokens::RawTokenAmount;
 use plt_token_module::token_module;
+use utils::kernel_stub::{KernelStub, TokenInitTestParams, TransactionExecutionTestImpl};
 
-mod kernel_stub;
 mod utils;
 
 /// Default protocol version used across the tests.
